requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
